Moving GithubUser to use yaml config. Working on #78
