+++ spec/DevBoard/Github/User/Entity/GithubUserRepositorySpec.php
@@ -1,5 +1,5 @@
 <?php
-namespace spec\DevBoard\Github\User;
+namespace spec\DevBoard\Github\User\Entity;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
@@ -10,7 +10,7 @@ class GithubUserRepositorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('DevBoard\Github\User\GithubUserRepository');
+        $this->shouldHaveType('DevBoard\Github\User\Entity\GithubUserRepository');
     }
 
     public function let(EntityManager $em, ClassMetadata $classMetadata)
