+++ src/DevBoard/Behat/Github/User/GithubUserContext.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Behat\Github\User;
 
 use Behat\Gherkin\Node\TableNode;
-use DevBoard\Github\User\GithubUser;
+use DevBoard\Github\User\Entity\GithubUser;
 use Resources\Behat\DomainContext;
 
 /**
