+++ src/DevBoard/Behat/NullDev/User/FixtureContext.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Behat\NullDev\User;
 
 use Behat\Gherkin\Node\TableNode;
-use DevBoard\Github\User\GithubUser;
+use DevBoard\Github\User\Entity\GithubUser;
 use Resources\Behat\DefaultContext;
 
 /**
