+++ src/DevBoard/Behat/NullDev/User/FrontContext.php
@@ -6,7 +6,7 @@ use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
 use Behat\MinkExtension\Context\RawMinkContext;
 use Behat\Symfony2Extension\Context\KernelDictionary;
-use DevBoard\Github\User\User;
+use DevBoard\Github\User\Entity\User;
 use Resources\Behat\DomainContext;
 use Resources\Behat\WebContext;
