+++ src/DevBoard/Github/User/Entity/GithubUser.php
@@ -1,89 +1,44 @@
 <?php
-namespace DevBoard\Github\User;
+namespace DevBoard\Github\User\Entity;
 
 use DateTime;
 use Doctrine\ORM\Mapping as ORM;
+use Resources\Entity\BaseEntity;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 
 /**
  * GithubUser.
  *
- * @ORM\Table(name="GithubUsers")
- * @ORM\Entity(repositoryClass="DevBoard\Github\User\GithubUserRepository")
  * @UniqueEntity("githubId")
  * @UniqueEntity("username")
  * @UniqueEntity("email")
- * @ORM\HasLifecycleCallbacks()
  */
-class GithubUser
+class GithubUser extends BaseEntity
 {
     /**
-     * @var int
-     *
-     * @ORM\Column(name="id", type="guid")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="UUID")
-     */
-    private $id;
-
-    /**
      * @var string
-     *
-     * @ORM\Column(name="githubId", type="string", length=255, nullable=true,unique=true)
      */
     private $githubId;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="username", type="string", length=255, nullable=true,unique=true)
      */
     private $username;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="email", type="string", length=255, nullable=true,unique=true)
      */
     private $email;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=255)
      */
     private $name;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="avatarUrl", type="string", nullable=true, length=255)
      */
     private $avatarUrl;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="createdAt", type="datetime")
-     */
-    private $createdAt;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="updatedAt", type="datetime")
-     */
-    private $updatedAt;
-
-    /**
-     * Get id.
-     *
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
 
     /**
      * Set githubId.
@@ -199,71 +154,6 @@ class GithubUser
     public function getAvatarUrl()
     {
         return $this->avatarUrl;
-    }
-
-    /**
-     * Set createdAt.
-     *
-     * @param \DateTime $createdAt
-     *
-     * @return GithubUser
-     */
-    public function setCreatedAt(DateTime $createdAt)
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    /**
-     * Get createdAt.
-     *
-     * @return \DateTime
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-    /**
-     * Set updatedAt.
-     *
-     * @param \DateTime $updatedAt
-     *
-     * @return GithubUser
-     */
-    public function setUpdatedAt(DateTime $updatedAt)
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
-    }
-
-    /**
-     * Get updatedAt.
-     *
-     * @return \DateTime
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-
-    /**
-     * @ORM\PrePersist
-     */
-    public function doCreatedValue()
-    {
-        $this->setCreatedAt(new \DateTime());
-        $this->setUpdatedAt(new \DateTime());
-    }
-
-    /**
-     * @ORM\PreUpdate
-     */
-    public function doUpdatedValue()
-    {
-        $this->setUpdatedAt(new \DateTime());
     }
 
     /**
